requirements.txt
Flask==1.1.1
Keras==2.2.4 # 2.3.0
# keras_tqdm==2.0.1
mrcnn==0.2
numpy==1.17.2
opencv-python==3.4.2.17
pandas==0.25.1
Pillow==6.2.0 
python-dotenv==0.10.3
tensorflow==1.14.0 # 2.0.0
matplotlib==3.1.1

#Seems that mrcnn have unmet dependencies 
cython
h5py
scikit-image
IPython

